updating api documentation

+++ backend/Electric/src/main/java/com/example/electric/service/UserService.java
@@ -1,20 +1,29 @@
 package com.example.electric.service;
 
+import com.example.electric.model.Role;
 import com.example.electric.model.User;
 import com.example.electric.respository.UserRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+
 @Service
-public class UserService {
+@RequiredArgsConstructor
+public class UserService{
+
+    @Autowired
+    private BCryptPasswordEncoder passwordEncoder;
 
     private final UserRepository userRepository;
 
     @Autowired
-    public UserService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository, BCryptPasswordEncoder passwordEncoder) {
+        this.passwordEncoder = passwordEncoder;
         this.userRepository = userRepository;
     }
 
@@ -28,21 +37,42 @@ public class UserService {
     }
 
     public User createUser(User user) {
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
+        user.setRole(Role.ROLE_USER);
         return userRepository.save(user);
     }
 
     public User updateUser(Long id, User updatedUser) {
         Optional<User> optionalUser = userRepository.findById(id);
+
         if (optionalUser.isPresent()) {
             User user = optionalUser.get();
             // Update the user fields as needed
-            user.setFirstName(updatedUser.getFirstName());
-            user.setLastName(updatedUser.getLastName());
-            user.setUsername(updatedUser.getUsername());
-            user.setEmail(updatedUser.getEmail());
-            user.setPassword(updatedUser.getPassword());
-            user.setCars(updatedUser.getCars());
-            user.setCard(updatedUser.getCard());
+            if(updatedUser.getFirstName() != null){
+                user.setFirstName(updatedUser.getFirstName());
+            }
+            if(updatedUser.getLastName() != null){
+             user.setLastName(updatedUser.getLastName());
+            }
+            if(updatedUser.getUsername() != null){
+            user.setUsernames(updatedUser.getUsername());
+            }
+            if(updatedUser.getEmail() != null){
+                user.setEmail(updatedUser.getEmail());
+            }
+            if(updatedUser.getPassword() != null){
+                user.setPassword(passwordEncoder.encode(updatedUser.getPassword()));
+            }
+            if(updatedUser.getCars() != null){
+                user.setCars(updatedUser.getCars());
+            }
+            if(updatedUser.getCard() != null){
+                user.setCard(updatedUser.getCard());
+            }
+//            if(updatedUser.getAuthorities() != null){
+//                user.setAuthorities(updatedUser.getAuthority());
+//            }
+
             return userRepository.save(user);
         } else {
             return null; // User not found
@@ -50,13 +80,21 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            return;
+        }
         userRepository.deleteById(id);
-    }
-    public long getUserIDByEmail(String email) {
-        return userRepository.getUserIDByEmail(email);
     }
 
     public User getUserByEmail(String email) {  // check if user exists
-        return userRepository.getUserByEmail(email);
+        Optional<User> optionalUser = userRepository.findByEmail(email);
+        return optionalUser.orElse(null);
     }
+
+    public boolean isEmailUnique(String email) {
+        Optional<User> existinguser = userRepository.findByEmail(email);
+        return !existinguser.isPresent();
+    }
+
 }
+
